feat(pku2u & pk_init): move pk-init related funtions into the pk_init module. refactor pku2u
